requirements.txt
beautifulsoup4>=4.12,<5
openai>=1.30,<2
python-dotenv>=1.0,<2
qdrant-client>=1.9,<2
requests>=2.31,<3
pytest>=8,<9
